retrieve user access token from auth flow
Previously we were only retrieving a temporary/exchange token using the Discord OAuth flow with AbandonAuth. I added an additional request that uses the exchange token to obtain a user access token and set the Authorization cookie to the long-lived user token. The site now has the auth required to manage developer applications for the authenticated user.